Remove static methods from RestClientHelpers to RestClientBuilder
