+++ src/AutoRest.CSharp/Output/Models/DataPlaneClient.cs
@@ -124,7 +124,7 @@ namespace AutoRest.CSharp.Output.Models
 
         public IReadOnlyCollection<Parameter> GetClientConstructorParameters(CSharpType credentialType)
         {
-            return RestClientHelpers.GetConstructorParameters(RestClient.Parameters, credentialType, false, false);
+            return RestClientBuilder<DataPlaneOutputLibrary>.GetConstructorParameters(RestClient.Parameters, credentialType, false, false);
         }
     }
 }
